refactor: separate Coroutine
